+++ CHANGES.rst
@@ -19,6 +19,10 @@ Changelog
 
 -   Use production accounting tag for bilby running on real events.
 
+-   Disable querying data before starting PE as bilby now uses kafka data
+    stream and PE timeout is long enough for low-latency frame data to be
+    produced.
+
 2.1.1 "Lone Island Mountain Devil" (2023-05-29)
 -----------------------------------------------